+++ packages/components/atoms/f-image-tile/src/components/ImageTile.vue
@@ -131,6 +131,15 @@ export default {
         isBreakoutImage: {
             type: Boolean,
             default: true
+        },
+        // An optional library for tracking rendering performance
+        performanceTracker: {
+            type: Object,
+            default: null
+        },
+        performanceTrackerLabel: {
+            type: String,
+            default: null
         }
     },
     data () {
@@ -164,6 +173,13 @@ export default {
     },
     created () {
         this.isToggleSelected = this.isSelected;
+    },
+    mounted () {
+        if (this.performanceTracker) {
+            this.$nextTick(() => {
+                this.performanceTracker.time(this.performanceTrackerLabel);
+            });
+        }
     },
     methods: {
         /**