+++ packages/components/molecules/f-restaurant-card/test/component/f-restaurant-card.component.spec.js
@@ -3,12 +3,12 @@ const RestaurantCard = require('../../test-utils/component-objects/f-restaurant-
 const restaurantCard = new RestaurantCard();
 
 describe('f-restaurant-card component tests', () => {
-    beforeEach(() => {
-        restaurantCard.load();
+    beforeEach(async () => {
+        await restaurantCard.load();
     });
 
-    it('should display the f-restaurant-card component', () => {
+    it('should display the f-restaurant-card component', async () => {
         // Assert
-        expect(restaurantCard.isComponentDisplayed()).toBe(true);
+        await expect(await restaurantCard.isComponentDisplayed()).toBe(true);
     });
 });