+++ packages/components/organisms/f-header/src/components/CountrySelector.vue
@@ -27,6 +27,9 @@
                 :class="$style['c-countrySelector-text']">
                 {{ copy.selectYourCountryText }}
             </span>
+            <chevron-right-icon
+                v-if="isBelowMid"
+                :class="$style['c-countrySelector-openPanelIcon']" />
         </button>
 
         <v-popover
@@ -48,11 +51,13 @@
 
 <script>
 import VPopover from '@justeat/f-popover';
+import { ChevronRightIcon } from '@justeattakeaway/pie-icons-vue';
 import CountrySelectorPanel from './CountrySelectorPanel.vue';
 import FlagIcon from './FlagIcon.vue';
 
 export default {
     components: {
+        ChevronRightIcon,
         CountrySelectorPanel,
         FlagIcon,
         VPopover
@@ -156,5 +161,11 @@ export default {
         margin: 0 spacing(d);
         float: left;
     }
+}
+
+.c-countrySelector-openPanelIcon {
+    float: right;
+
+    height: 18px;
 }
 </style>