+++ packages/components/pages/f-checkout/src/store/checkout.module.js
@@ -249,8 +249,12 @@ export default {
          */
         getNotesConfiguration: async ({ commit, state }, { url, timeout }) => {
             const { data } = await checkoutApi.getNoteConfiguration(url, timeout);
-            const isSplitNotesEnabled = !data?.customerNotes?.serviceTypes[state.serviceType].orderNoteAccepted;
-            commit(UPDATE_NOTES_CONFIGURATION, { ...data?.customerNotes?.serviceTypes, isSplitNotesEnabled });
+            if (data?.customerNotes.serviceTypes) {
+                // This has to be done because the dinein key is sent to us as 'dineIn` ....
+                const notesByServiceType = Object.fromEntries(Object.entries(data.customerNotes.serviceTypes).map(([k, v]) => [k.toLowerCase(), v]));
+                const isSplitNotesEnabled = !notesByServiceType[state.serviceType].orderNoteAccepted;
+                commit(UPDATE_NOTES_CONFIGURATION, { ...notesByServiceType, isSplitNotesEnabled });
+            }
         },
 
 
@@ -323,7 +327,6 @@ export default {
                 },
                 timeout
             };
-
             const { data } = await axios.get(url, config);
 
             const addressDetails = addressService.getClosestAddress(data, tenant, currentPostcode);