+++ packages/components/pages/f-checkout/stories/helpers/customer/ie.js
@@ -1,0 +1,7 @@
+export default {
+    firstName: 'Joe',
+    lastName: 'Bloggs',
+    phoneNumber: '+35371111111111',
+    dateOfBirth: '1980-01-30',
+    emailAddress: 'joe.bloggs@justeattakeaway.com'
+};