+++ packages/components/atoms/f-button/CHANGELOG.md
@@ -3,6 +3,26 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+Latest (add to next release)
+------------------------------
+*May 26, 2021*
+
+### Changed
+- Refactor WebDriverIO tests to use async in order to support Node 16 using `codemod` utility.
+
+
+v3.4.0
+------------------------------
+*May 13, 2022*
+
+### Changed
+- Icons in component and storybook files
+
+### Removed
+- `@justeat/f-vue-icons`
+
+### Added
+- `@justeattakeaway/pie-icons-vue`
 
 
 v3.3.1