+++ packages/components/molecules/f-mega-modal/CHANGELOG.md
@@ -3,6 +3,28 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+Latest (add to next release)
+------------------------------
+*May 26, 2021*
+
+### Changed
+- Refactor WebDriverIO tests to use async in order to support Node 16 using `codemod` utility.
+
+
+v5.1.0
+------------------------------
+*May 13, 2022*
+
+### Changed
+- Icons in component
+
+### Removed
+- `@justeat/f-vue-icons`
+
+### Added
+- `@justeattakeaway/pie-icons-vue`
+
+
 v5.0.0
 ------------------------------
 *April 11, 2022*