+++ packages/components/molecules/f-navigation-links/CHANGELOG.md
@@ -3,6 +3,21 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+Latest (add to next release)
+------------------------------
+*May 26, 2021*
+
+### Changed
+- Refactor WebDriverIO tests to use async in order to support Node 16 using `codemod` utility.
+
+
+v1.2.0
+------------------------------
+*May 22, 2022*
+
+### Added
+- Version of `node-sass` that's compatible with Node 16.
+
 v1.1.0
 ------------------------------
 *March 15, 2022*