+++ packages/services/f-services/CHANGELOG.md
@@ -3,6 +3,22 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v1.16.0
+------------------------------
+*May 13, 2022*
+
+### Added
+- Irish phone number validations
+
+
+v1.15.0
+------------------------------
+*May 11, 2022*
+
+### Added
+- Irish postcode validations
+
+
 v1.14.0
 ------------------------------
 *February 21, 2022*